revert package renaming
