+++ src/main/java/br/com/zup/kafka/config/props/core/GenericBuilder.java
@@ -1,4 +1,4 @@
-package br.com.zup.kafka.consumer.config.props.core;
+package br.com.zup.kafka.config.props.core;
 
 import java.util.Properties;
