+++ src/main/java/br/com/zup/kafka/consumer/ConsumerRunner.java
@@ -1,12 +1,12 @@
-package br.com.zup.kafka.consumer.consumer;
+package br.com.zup.kafka.consumer;
+
+import br.com.zup.kafka.config.props.ConsumerProperties;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
-
-import br.com.zup.kafka.consumer.config.props.ConsumerProperties;
 
 public class ConsumerRunner {
