+++ src/main/java/br/com/zup/kafka/consumer/config/KMessageConsumer.java
@@ -1,6 +1,6 @@
-package br.com.zup.kafka.consumer.consumer.config;
+package br.com.zup.kafka.consumer.config;
 
-import br.com.zup.kafka.consumer.KafkaMessage;
+import br.com.zup.kafka.KafkaMessage;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 
 public interface KMessageConsumer<K, V> {
