+++ src/main/java/br/com/zup/kafka/producer/KafkaProducer.java
@@ -1,15 +1,15 @@
-package br.com.zup.kafka.consumer.producer;
+package br.com.zup.kafka.producer;
 
-import java.util.Map;
-import java.util.concurrent.Future;
-
-import br.com.zup.kafka.consumer.KafkaMessage;
-import br.com.zup.kafka.consumer.config.props.ProducerProperties;
+import br.com.zup.kafka.KafkaMessage;
+import br.com.zup.kafka.config.props.ProducerProperties;
 import org.apache.kafka.clients.producer.Producer;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.clients.producer.RecordMetadata;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.util.Map;
+import java.util.concurrent.Future;
 
 public class KafkaProducer<K, V> {
