+++ src/main/java/br/com/zup/kafka/producer/serializer/JsonSerializer.java
@@ -1,14 +1,13 @@
-package br.com.zup.kafka.consumer.producer.serializer;
+package br.com.zup.kafka.producer.serializer;
 
-import java.io.UnsupportedEncodingException;
-import java.util.Map;
-
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import org.apache.kafka.common.errors.SerializationException;
 import org.apache.kafka.common.serialization.Serializer;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import java.io.UnsupportedEncodingException;
+import java.util.Map;
 
 public class JsonSerializer implements Serializer<Object> {
